Folder structure edit
