--- packages/storefront/components/checkout/index.tsx
@@ -1,17 +1,0 @@
-import InformationPage from "./informationPage";
-import OrderList from "./orderList";
-
-const CheckoutComponent = (props: any) => {
-  return (
-    <>
-      <div className="row">
-        <div className="divide-x-0 sm:divide-x-0 md:divide-x-0 lg:divide-x-2 xl:divide-x-2 flex flex-wrap flex-col-reverse sm:flex-col-reverse md:flex-col-reverse lg:flex-row xl:flex-row justify-between">
-          <InformationPage/>
-          <OrderList />
-        </div>
-      </div>
-    </>
-  );
-};
-
-export default CheckoutComponent;
