--- packages/storefront/pages/checkout.tsx
@@ -1,11 +1,0 @@
-import type { NextPage } from "next";
-import CheckoutComponent from "../components/checkout";
-
-const Checkout: NextPage = () => {
-  return (
-    <>
-        <CheckoutComponent />
-    </>
-  );
-};
-export default Checkout;
