+++ packages/storefront/pages/checkoutInformation.tsx
@@ -1,0 +1,11 @@
+import type { NextPage } from "next";
+import CheckoutInformationComponent from "../components/checkout/informationPage/information";
+
+const CheckoutInformation: NextPage = () => {
+  return (
+    <>
+        <CheckoutInformationComponent />
+    </>
+  );
+};
+export default CheckoutInformation;
